Swap out fledgling Flask application for a Node Express one,
Investigation revealed that Web3.py couldn't read the events raise
by our smart contract. This was, apparently, a problem in the Solidity
compiler rather than Web3.py, but that's little consolation as it means
I can't use it.  Switching to a JavaScript backend because we know
Web3.js works ok, even though it means we'll have to shell out to
run the image fingerprinter.